refactor: use headers with minimal overhead
