+++ src/core/encryption/ciphers/Serpent/index.ts
@@ -1,8 +1,24 @@
 import { init, SerpentCtr } from 'leviathan-crypto';
 import { IEncryptionProcessor, RandomBytesGenerator } from '@core/encryption';
 import { fillBuffer, joinBuffers } from '@core/encryption/utils/buffers';
+import { bytes, struct, u8 } from '@core/encryption/utils/bytes/binstruct';
 
-import { transformBuffer, TwofishBufferHeader } from '../Twofish';
+/**
+ * Inner util to transform block with cipher
+ * Util creates a new buffer instead of mutate original buffer.
+ */
+export function transformBuffer(
+	buffer: Uint8Array,
+	transformBlock: (offset: number, input: Uint8Array, output: Uint8Array) => void,
+	blockSize = 16,
+) {
+	const out = new Uint8Array(buffer.length);
+	for (let offset = 0; offset < buffer.length; offset += blockSize) {
+		transformBlock(offset, buffer, out);
+	}
+
+	return out;
+}
 
 let initStatus: boolean | Promise<void>;
 export async function ensureWasmIsLoaded() {
@@ -17,7 +33,11 @@ export async function ensureWasmIsLoaded() {
 	await initStatus;
 }
 
-const binHeader = new TwofishBufferHeader(16);
+export const SERPENT_IV_SIZE = 16;
+export const SERPENT_HEADER = struct({
+	padding: u8(),
+	iv: bytes(SERPENT_IV_SIZE),
+});
 
 const chunkSize = 65536;
 
@@ -30,12 +50,11 @@ export class SeprentCipher implements IEncryptionProcessor {
 	public async encrypt(buffer: ArrayBuffer) {
 		await ensureWasmIsLoaded();
 
+		const iv = this.getRandomBytes(SERPENT_IV_SIZE);
 		const [bufferView, padding] = fillBuffer(new Uint8Array(buffer));
 
-		const nonce = this.getRandomBytes(16).buffer;
-
 		const cipher = new SerpentCtr({ dangerUnauthenticated: true });
-		cipher.beginEncrypt(this.key, new Uint8Array(nonce));
+		cipher.beginEncrypt(this.key, iv);
 		const encryptedBuffer = transformBuffer(
 			bufferView,
 			(offset, input, output) => {
@@ -48,23 +67,21 @@ export class SeprentCipher implements IEncryptionProcessor {
 		);
 		cipher.dispose();
 
-		return joinBuffers([
-			binHeader.createBuffer({ iv: nonce, padding }),
-			encryptedBuffer,
-		]);
+		return joinBuffers([SERPENT_HEADER.encode({ padding, iv }), encryptedBuffer]);
 	}
 
 	public async decrypt(buffer: ArrayBuffer) {
 		await ensureWasmIsLoaded();
 
-		const { iv, padding } = binHeader.readBuffer(
-			buffer.slice(0, binHeader.bufferSize),
+		const bufferView = new Uint8Array(buffer);
+		const { padding, iv } = SERPENT_HEADER.decode(
+			bufferView.subarray(0, SERPENT_HEADER.size),
 		);
 
 		const cipher = new SerpentCtr({ dangerUnauthenticated: true });
-		cipher.beginEncrypt(this.key, new Uint8Array(iv));
+		cipher.beginEncrypt(this.key, iv);
 		const decryptedBuffer = transformBuffer(
-			new Uint8Array(buffer.slice(binHeader.bufferSize)),
+			bufferView.subarray(SERPENT_HEADER.size),
 			(offset, input, output) => {
 				output.set(
 					cipher.decryptChunk(input.slice(offset, offset + chunkSize)),
